fix: CG-H9 sprintf returns i32 (was i64)
sprintf in C returns int (i32), not i64. Fixes LLVM function type
mismatch that could cause issues with optimization and linking.
Golden IR files updated to reflect the new signature.

不变量类别: L1
测试: 2900 passed, 0 failed, 7 ignored

+++ src/codegen/builtins/mod.rs
@@ -180,7 +180,7 @@ fn register_libc<'ctx>(
 
     module.add_function(
         "sprintf",
-        i64.fn_type(
+        i32.fn_type(
             &[
                 BasicMetadataTypeEnum::PointerType(i8_ptr),
                 BasicMetadataTypeEnum::PointerType(i8_ptr),
